Message_push

Closes #13

See merge request internship/upmureport!49

+++ src/main/java/com/nastech/upmureport/domain/repository/ProjectRepository.java
@@ -1,5 +1,7 @@
 package com.nastech.upmureport.domain.repository;
 
+import java.math.BigInteger;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.transaction.annotation.Transactional;
@@ -7,5 +9,5 @@ import org.springframework.transaction.annotation.Transactional;
 import com.nastech.upmureport.domain.entity.Project;
 
 @Transactional
-public interface ProjectRepository extends JpaRepository<Project, Integer>{
+public interface ProjectRepository extends JpaRepository<Project, BigInteger>{
 }
